- Added: list command - Added: listadd command - Added: listdel command
list command is used to get the list of inappropriate bannable and non
bannable usernames. Bannable list is DM-ed to the user and the other
list is shown in the channel. listadd is used to add another word into
one of the lists. listdel is used to remove a word from one of the lists.